refactor(suite): unify the Mains
